Microcommerce

+++ Mcommerce/Microservice-paiement/src/main/java/com/example/microservicepaiement/model/Paiement.java
@@ -17,7 +17,7 @@ public class Paiement {
     @Column(name = "id_commande", unique = true)
     private Integer idCommande;
 
-    private Integer numeroCarte;
+    private Long numeroCarte;
 
 
 }
